Handle promise error on http sites; http perf improvement
In Chrome 42, an error is logged when a promise has an error that is
not handled. So, even though this block has a try-catch around it, we
need to handle the case where the site is “non-secure” (http) and
supports crypto.subtle, in which case this first promise will be
rejected.

This error:

Uncaught (in promise)
DOMException: Only secure origins are allowed. http://goo.gl/lq4gCo

While we’re at it, let’s set `subtle = false` so that after the first
failed promise we don’t need to pay the cost of failed promises for
each call to sha1()